wptraffic/requirements.txt
requests==2.31.0
python-wordpress-xmlrpc==2.3
faker==22.0.0
